Move CreateAccountInputs to a separate file
